refactor: Extract PeerConnectionInfo and config helper method in its own module
