+++ daemon/src/lib.rs
@@ -3,6 +3,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+pub mod config;
 pub mod daemon;
 pub mod document;
 pub mod editor;
